rename web3 files
